requirements.txt
fastapi
uvicorn
pandas
numpy
textblob
waitress
pymongo
openai
playwright
python-dotenv
